pyproject.toml
[project]
name = "hw-verify-probe"
version = "0.1.0"
description = "Hardware Verification & Probing Tool for MacOS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.2",
    "grpcio>=1.80.0",
    "grpcio-tools>=1.80.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.1",
    "pyserial>=3.5",
    "pyvisa>=1.16.2",
    "pyyaml>=6.0.3",
]

[project.scripts]
# 修改入口命令，以后在终端输入 hwv 即可启动
hwv = "main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# 显式包含根目录下的 .py 文件以及 app, devices, engine, utils 文件夹
packages = [
    "app",
    "devices",
    "engine",
    "utils",
    "main.py"
]

# 如果你的协议类文件在根目录，也需要包含进去
include = [
    "/*.py",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.19.0",
]
